So many things done... Really not following BPs with this one. Updated all of the company views, they now look nice. Database updated massively and stored locally on PC or in android emulator data/data/com.example.investic/databases folder.
CompanyArticle objects set up for future use.

DBHelper has loads of new methods, primarily around company details and profile/portfolio stuff.

PortfolioAdapter constructs cardviews for each of the companies in the portfolio and adds them into the Portfolio Fragment.

CompanyListAdapter does the same, but for the Home Fragment. This is where the full list of companies is currently visible.

The HomeFragment company list can be searched using an edit text box, which passes the entry into a method in the main activity, which passes it into an instance of the DBHelper, which then performs the query and returns an arraylist of company objects.